requirements.txt
flask
pandas
gspread
google-auth
google-auth-oauthlib
